Adicionado API para contabilizar votos

+++ src/main/java/com/github/ricardorv/desafiosicredi/api/v1/dto/ResultadoVotacaoDto.java
@@ -15,7 +15,7 @@ import java.util.Map;
 @AllArgsConstructor
 public class ResultadoVotacaoDto {
 
-    Map<VotoEnum, Integer> quantidadeVotos;
+    Map<VotoEnum, Long> quantidadeVotos;
     LocalDateTime encerramento;
 
 }
